Move FragmentId into android package.
