+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Rename `KeyState` to `FragmentState` and move it to `android` package.
 - Move `ActivityStoreContext` into `android` package.
 - Move `ActivityStore` into `android` package.
+- Move `FragmentId` into `android` package.
 
 ## [0.6.1] - November 18, 2020
 - Bugfix: Fix runtime ignoring `Formula.key` for the root formula.
