+++ formula-android/src/main/java/com/instacart/formula/android/internal/ActivityStoreContextImpl.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.Lifecycle
 import com.instacart.formula.activity.ActivityResult
 import com.instacart.formula.fragment.FragmentFlowState
 import com.instacart.formula.fragment.FragmentKey
-import com.instacart.formula.integration.FragmentId
+import com.instacart.formula.android.FragmentId
 import com.instacart.formula.android.ActivityStoreContext
 import com.jakewharton.rxrelay3.BehaviorRelay
 import com.jakewharton.rxrelay3.PublishRelay
