+++ formula-android/src/main/java/com/instacart/formula/android/internal/FeatureProvider.kt
@@ -1,7 +1,7 @@
 package com.instacart.formula.android.internal
 
 import com.instacart.formula.integration.FeatureEvent
-import com.instacart.formula.integration.FragmentId
+import com.instacart.formula.android.FragmentId
 
 internal interface FeatureProvider {
     fun getFeature(id: FragmentId): FeatureEvent?
