+++ formula-android/src/main/java/com/instacart/formula/integration/Binding.kt
@@ -2,6 +2,7 @@ package com.instacart.formula.integration
 
 import com.instacart.formula.FormulaContext
 import com.instacart.formula.android.FlowFactory
+import com.instacart.formula.android.FragmentId
 import com.instacart.formula.android.internal.ComponentFactory
 import com.instacart.formula.fragment.FragmentEnvironment
 import com.instacart.formula.android.internal.CompositeBinding
